+++ abc/abc151-200/src/abc194/abc194F.cpp
@@ -9,7 +9,7 @@ const vector<char> NUMBERS = { '0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
 
 int main() {
 	string n;
-	long unsigned int k;
+	int k;
 	cin >> n >> k;
 	int l = n.size();
 	vector<vector<mint>> dp(l + 1, vector<mint>(N + 1));
@@ -27,12 +27,12 @@ int main() {
 		if (i > 0) {
 			dp[i + 1][1] += N - 1;
 		}
-		for (long unsigned int j = 1; j <= k; j++) {
+		for (int j = 1; j <= k; j++) {
 			dp[i + 1][j] += dp[i][j] * j;
 			dp[i + 1][j] += dp[i][j - 1] * (N - (j - 1));
 		}
 	}
-	if (k == s.size()) {
+	if (k == int(s.size())) {
 		dp[l][k]++;
 	}
 	cout << dp[l][k].val() << endl;
